Include 'council' in model and make sure warning are correct
Currently the popolo generated includes a warning about skipping the
'council' column. This is a lie, because we are using that column to
create popolo data.

It thinks it is skipping it because it is not in the model.

Add 'council' to the model so that the warning reflect reality.