There is no independent executive organisation for councils, don't create one
Currently the converter assumes we want to create a default independent
executive organization. For our councillor data, we don't have this.

Councillors can have executive posts at the council organizations they
are also councillors at. We're capturing this as two separate
memberships with different roles at the same organization.

This commit removes the call the create executive organisations if their
are councillors in the csv.